[INTERNAL][FIX] sap.f.Avatar: Focus outline is fixed
Problem: Focus outline of Avatar control is not visible on all browsers,
except Internet Explorer (because it does not support outline-offset).

Solution: Outline-offset is not set to -1px now, because this way it
overlaps Avatar and is not clearly visible when the Image is in a
similar color.

Change-Id: Idc225c7faf244967cd8fa94465ecdf7931b6698a
JIRA: BGSOFUIPIRIN-3699