[FIX] sap.m.SuggestionsPopover: Cursor now stays on its correct place
After there were no found suggestions for a text,
the cursor was always moving to the end of the input.
Now it stays where it was.

The oposite was happening when a suggestion was clicked -
the cursor did stay where it was, while it should go to the end.
This is now fixed as well.

Change-Id: I0599ad25e88c8a2f387b8fc2153230df08ee720a
BCP: 1970127503
Snippix: 99972