[FIX] sap.uxap.ObjectPageSection: New title of section is shown on reveal
Problem: When there is only one ObjectPageSubSection, its title is shown
in the AnchorBar, instead of the title of the ObjectPageSection itself.
When the title of the SubSection is changed, then the SubSection is
deleted and after that we are trying to restore back the whole Section,
the old title of the Section is shown, because actually the SubSection
has been renamed. This case is misleading for users because they have
the feeling that they renamed the Section, not the SubSection (as they
can see only the title of the SubSection).

Solution: A new getLabel function is implemented in the
ObjectPageSection.designtime.js. When the ObjectPageSection has only one
SubSection, the title of the SubSection is returned, instead of the
title of the Section.

Change-Id: Ieb8d3e7b1206903c2f4f7e1c3556b9b60b92f904
BCP: 1970093334