[FEATURE] sap.m.SinglePlanningCalendar: Appointment all-day is reworked
All-day appointment is an appointment which starts at 00:00 and
ends in 00:00 on any day in the future.
This change removes the experimental sap.m.CalendarAppointment and its
fullDay property, reuses the sap.ui.unified.CalendarAppointment and adds
an all-day calculation.
This is represented in a Demo Kit with a checkbox for switching a
DateTimePicker or DatePicker, depending on whether or not it is an
all-day appointment.

Change-Id: I2f4e6634e4c35c8149ddb705d9046fa20d45ef4a
JIRA: BGSOFUIBALKAN-4590