[INTERNAL] caja-html-sanitizer: wrap in IIFE to avoid top level vars
Without further configuration, modules that define top level variables
will be packaged as strings by the bundle tooling and eval'd at runtime.
To avoid this for the often-used caja-html-sanitizer, the whole module
is wrapped in an IIFE. This module exports its APIs to the window object
anyhow.

Avoiding the eval call improves compliance with CSP target level 2.

Change-Id: I29a8099246a153875a21c533d6aee164e1f710cb