[FEATURE] sap.uxap.ObjectPageDynamicHeaderTitle: Single-line title on phone
Description:

The only content that is displayed in the DynamicPageTitle
when it is viewed on a phone mobile device and the ObjectPageHeader
is in collapsed (snapped) state.

Using this aggregation enables you to provide a simple, single-line
title that takes less space on the smaller phone screens when the
ObjectPageHeader is in its collapsed (snapped) state.

Change-Id: I30b2896839343aeae87a02098c6f294cc82546f7
JIRA: BGSOFUIPIRIN-3514