[FIX] Demo Kit: Samples navigation and deep links fixed
The problems:
1. Navigation back, previous and next sample on a sample which is
attached to multiple entities fails as the back navigation is pointing
towards the last entity claiming the sample.
2. Deep link to a sample on phone - back button is hidden.
3. When previewing a file content from a sample, the entity to which the
sample belongs is not selected in the master list.
4. On deep link to a sample, the entity to which the sample belongs is
not selected on the master list.

The root cause:
All mentioned problems occur because of are a consequence of:
Code relied that one sample belonged to just one entity and that on
mobile there would be no back navigation if no history was present.

The solutions:
* Introducing new URL format for samples containing context from which
example the control is opened for all three views ("sample", "code",
"codeFile"). This way, on all mentioned pages we have a context from
which entity the sample was opened from and we can appropriately setup
"back", "next" or "previous" sample navigation even on deep link. This
includes collecting context data for each sample in ControlsInfo class.

* Legacy redirects are added for all three old format URL's which don't
contain context (entity) info. In this case we redirect to a new style
URL and for entity we assign the first context that ControlsInfo returns
for the sample.

* Modifyed ControlsMaster.controller to handle new routes and always
have a selected entity.

* Minor reformatting of manifest.json.

Change-Id: I9b494f39081e38b40fe8b504e1283b888ec82c00
BCP: 1970068289