[FIX] sap.m.DatePicker: year picker now properly honors minDate.
Issue:
If the apps set a min date, year picker does not fully respect it for
browsers with negative timezones and does not allow choosing the min
possible year in the year picker. The picker disabled the "backward"
arrow button too early - e.g. if min date is 1 of Jan 2008, the first
visible year in the picker is 2009, there should be possible to navigate
backward, but the mentioned button is disabled.

Caused by:
bug in YearPicker.prototype.getFirstRenderedYear() where the returned
JS Date object was not created by means of local JS Date constructor.

BCP: 1970162005

Change-Id: I9e11f5018260148caa7b02fa881e9d0bb1506a6b
Closes: https://github.com/SAP/openui5/issues/2424