[INTERNAL][FIX] sap.m.Panel: Improve Accessibility in Chrome
JIRA: BGSOFUIRILA-2348

This change fixes an issue in Chrome and FF, where the title
of the panel is announced twice, when an expandable panel button is on focus.
For IE and Edge the aria-labelledby is needed since the speech
output does not include information regarding the panel,
when the expand button is on focus.

Change-Id: Ic6e640e06ff341613715336908bb1fad8a0005ea