[FEATURE] sap.f.ShellBar: Created notifications badge element
Inside the Fiori 3 ShellBar header, where we can find notifications
button, we now show additional element that visualizes the
number of notifications for the application.

When the resize breakpoint is hit and the button enters in
OverflowToolbar the element is shown at the top right corner
of the OverflowToolbarButton. Along with that, we still
show it attached to the notification button when inside the
Overflow area.

Change-Id: Id512c1349e730d11bf78613d3bbb319b4972b178
JIRA: BGSOFUIPIRIN-3717