[FIX] ManagedObjectModel: Paging
Note:
In case a list/table was bound to a managed object items where the
length of these items is growing the MOM returned no item at least.
Ideally it should trigger a paging in the inner binding.
Which means when knowing there is an inner binding ideally the length
of this binding should be returned to a list or table.

Note 2:
We can assume as for now to only determine the oInnerBinding in case the
ManagedObjectModel binding points directly to a member.
For all other parts we have a JSON structure to assume. In case the
binding of the root object is an oData binding all the other parts go
via auto expand of navigation properties which is at least in oData v2
only possible to a certain level and there no $top&$skip are used.
For OData v4 we maybe have to investigate.

Note 3:
This fix also repairs the extended change detection, see getEntryKey.
Background: Due to the possibility of are properties
(see mdc.Field conditions) there may be composite visualizations that
use the array property as list for rendering the items. Thus the
oObject.getContext() not in any case a managed object but kan be a plain
JSON object also.

Change-Id: I6d38b14ec3b17affc13dbd320e57308f6ff67afa