[FIX] sap.ui.layout.BlockLayoutCell: Width of the cell is now correct
BCP: n/a

Problem: When the width from binding was set it not affect the cell
and it stays default. This come from the fact that when
setWidth method was called the layout data was not created or if
it was created the values for the different sizes was not updated.

Change-Id: Ia01d0b95e03ce5673dfe7b3e60f68120692ff39d