[INTERNAL][FIX] sap.uxap.AnchorBar: Updated Breakpoint classes in Fiori 3
Previously: The Fiori 3 theme was using the old/incorrect "sapUiMedia"
breakpoint CSS classes, which are aligning the content according to the
browser window width.

Now: The Fiori 3 theme is using the new/correct "sapUxAPObjectPageLayout"
breakpoint CSS classes, which are aligning the content according to the
width of the ObjectPage in which the AnchorBar is being used.

Change-Id: I96ee2a65ba48aed3588b3b58753a151eaa6f0c62