[FIX] sap.f.Avatar: Failing QUnit on IE is fixed
Problem: Internet Explorer does not support outline-offset and checking
for not-existing css property returns undefined.

Solution: Check for truthy statement - the outline-offset value should
be either "0px" or undefined. After all, the main purpose of this QUnit
is to ensure the outline-offset property is not set when the Avatar is
focused, because previously the focus outline appeared on the Avatar,
which in some cases made it invisible (or hardly visible).

Change-Id: If8078c4d997e2e43dbc8ab14c14b7151a72bf903
BCP: 1980015035