[FIX] sap.f.FlexibleColumnLayout: Fixed column resizing on layout change
Problem: Changing the layout of sap.f.FlexibleColumnLaout from
sap.f.LayoutType.TwoColumnsMidExpanded to
sap.f.LayoutType.EndColumnFullScreen causes the endColumn
to be shown only on half of the screen.

Root cause: The variable, containing the new width of a column, was
declared outside of the scope of the loop that calculates the new width.
This variable is used on a later point (when all animations are over)
to set the new width of a column. Since we have iterated all columns the
value in the variable contains the new width of the last column
instead of the value of the new width for the corresponding column.

Solution:
Move the variable to the scope of the loop that calculates the new width
of the columns.

Change-Id: I11b6d9406187fa7c3938d2e3b1ef7ea84c0375fb
CR-Id: 002075125800000457682019
BCP: 1970178100