[INTERNAL][FIX] sap.ui.fl.ChangePersistence resets personalization
With the introduction of Change-Id:
Ib07326698d123370d4fad108f319e642a588937c the changePersistence does no
longer check for transports in the case of a reset of personalization.
Thus a property aChanges is not defined as an array and may break a
.some call.
This variable is now defined accordingly. + test to ensure this
behavior.

Change-Id: Ic40bd6f67f4a7f86a5f45e9ffbd3c1dfef6a5716