[FEATURE] Messaging: Introduce MessageScope
With this change the business object message handling is introduced. If
the server signals with the annotation sap:message-scope-supported="true" that
message scopes are supported and the ODataModel message scope is set to
"Business object", the root entity is send with the "sap-message-scope" header
in every OData request. With this header the server is able to send only relevant
error messages.

JIRA: CPOUIFPHOENIXCORE-2737

Change-Id: I7066adc665f43252d288ec95104a64243041148c