[INTERNAL] Table: Make column header cell content accessible with F2
Column header cells are not part of the edit mode navigation.
But there should be a way to access their interactive content with the
keyboard.

The F2 keyboard shortcut was enhanced to be able to enter and leave
interactive column header cell content. Furthermore, pressing TAB or
Shift+TAB moves the focus back to the column header cell (see the first
sentence) and does not throw an error anymore.

Change-Id: I8bfce36c5b07e54c1dba64f34dd2bc76ae9598ea
BCP: 1980008380