[FEATURE] sap.f.DynamicPageTitle: snapped/expanded content redesign
Up to now, the snappedContent and expandedContent aggregations of
sap.f.DynamicPageTitle had the limited width of the heading aggregation.
This leads to visual degradations and low flexbility of the snapped and
expanded parts.

Now, the snappedContent and expandedContent aggregations are redesigned
to take the whole width of the title are no longer limited to the
heading width.

Change-Id: I68671588106f70d4034246b547953b3cc3afa536
JIRA: BGSOFUIPIRIN-3756