[INTERNAL][FEATURE] sap.m.Input: Enable Popover w/ Items for input controls
For the ML scenarios, we need a way to enable suggestion items
for the input controls.
Sap.m library is too generic to cover just a single scenario and we needed
to find a generic way to enable the ML use case.

This change enables app developers to show Popover for certain input
controls and filter the available items on custom criteria.

JIRA: BGSOFUIRILA-2357

Change-Id: I4398ca4abb5324ce1d2a3d128019faa37b3e1d43