[INTERNAL] v2.ODataModel: Fix another issue with updating nav properties in changed entities
When automatic refresh of entities is done after updates, the refresh request does not expand any
navigation properties. So the response contains __deferred for all navaigation properties. When
importing this response in the model data, deferred navigation properties do not overwrite
existing navigation properties, so the old navigation property value stays in the model cache
data. When updating the changed entities after import, the changed property value is removed from
the changed entity, but the corresponding navigation property does not match the current (old)
model value, so it stays in the changed entity. Next time submitChanges is called, this was
causing a MERGE request without data.

This change takes care to not only remove the navigation property from the changed entity data,
when the property itself is removed (to avoid the empty MERGE request), but also sets the new
navigation property value in the model cache data, so it does match the updated property value.

Change-Id: Ic37394275cb976ab8fb811c6093cb76ba7db1a66
BCP: 1980017086