[FIX] sap.ui.unified.CalendarTimeInterval: Navigates to the correct date
If in a PlanningCalendar (Hours view) Jan 31st was selected from the Calendar,
and then if you select again from the Calendar 1st of February,
the date was changed to March 1st. Now it focuses Feb 1st
and that's applicable to all months which are following a month with 31 days.

This is caused because of "setUTCMonth" and JavaScript's understanding regarding
Gregorian calendars with 31 days.
 Example: December 31st is selected. If the user sets the month to November,
 since November 31st does not exists, JavaScript gives December 1st.

BCP: 0020751295 0000061772 2019

Change-Id: I59f53d4c14f29fe347fbcdc50c9afee0c6d6aeea
(cherry picked from commit fbf7dbd003d714da7fd8340cc72f76c2864ad051)
CR-Id: 002075125800000827282019