[FIX] sap.uxap.ObjectPageLayout: Corrected lazyLoading memory leak
Background:
A helper object for lazyLoading is created in the onBeforeRendering()
lifecycle method of the ObjectPage.
The helper object maintains an interval task for lazyLoading

Problem:
A new helper is created in each call of the onBeforeRendering()
lifecycle method =>
the old object and its timer task remained in memory

Solution:
Implemened 'destroy' on the helper that stops any running
interval task so it can be garbage collected

Note:
A next step (also tested) would be to refactor the existing
code to *reuse the same helper* on each rendering. Decided to
consider such further optimization in a separate change.

Change-Id: I076c972242a6c239f941e82f78d42930245e1916
BCP: 1780437211