[FIX] sap.f.FlexibleColumnLayout: Fixed fullscreen button in end column
Problem:
Clicking the fullscreen button in the end column does not switch to
the correct layout.

Root cause:
There was a new parameter introduced in the "detailDetail" route.
The functions for handling the fullscreen functionality in the end
column do not pass the new parameter to the router.

Solution:
Pass the missing parameter to the router in the functions
that handle the fullscreen functionality for the end column.

Change-Id: I7f50600cd91e0498cf16ca65106d769b163d1967
BCP: 1980052686