[FIX] AlignedFlowLayout: Overlapping issue of items fixed
The issue:
Incorrect arrangement of items across the available horizontal space
when the AlignedFlowLayout GUI control is initially rendered and an
ancestor DOM element is hidden (the CSS "display" style property is
set to "none").

The root cause:
Previously to this change, the width of the last invisible spacer was
only updated on after rendering. However, in some scenarios, for example,
when an ancestor DOM element is hidden, the width for the last spacer was
incorrectly calculated.

The solution:
Calculate and update the width for the last invisible spacer item on
content resize and theme changes.

Change-Id: Ib0ae29a2386faef1c624fc53eb4de8ce8a42ddaa
BCP: 1980003456