[FEATURE] sap.m.Select: Display multiline value state text
The sap.m.Select control is now enabled to display the text set in the
valueStateText property on multiple lines when the dropdown is
opened. This ensures that the value states are shown, even though the
text is longer than the sap.m.Select container itself.

To do this, we provided a new private control -
sap.ui.layout.SimpleFixFlex. The new control consist of two major
blocks - a fix content and a flex content.
The flex content adapts it's size to the fix content in order to
fill the empty space in the parent container.

The sap.m.Select uses the fix content of the sap.ui.layout.SimpleFixFlex
control in order to display the text set in the valueStateText property
and the flex container in order to display the sap.m.SelectList.

JIRA: BGSOFUIPIRIN-3623
Change-Id: I56bdb5d3b061820840d37a6b3b4d131a6b234eb6