[FEATURE] ManagedObjectObserver: Observe parent change
JIRA: CPOUIFDENVER-1272

Note:
It is now possible to observe a parent change. The mutations for this
parent change are 'set' for setting the parent and 'unset' for clearing
a parent (setting it to null).
Since both, the removal and the add of aggregations invoke a parent
change now also the the observer gets notified for the child here.

Change-Id: Ib652a4bc49ed6a9dd889513134874a817d80bf8b