[INTERNAL][FIX] sap.m.Carousel: Active pages are now correctly udpated
Problem: When Carousel is not rendered and the setActivePage method is
called after rendering of Carousel control, the active pages collection
is not correct because it has not been updated. This leads to problems
when the page is changed after that and Carousel's events are fired with
wrong collection.

Solution: Update the active pages in onAfterRendering method.

Change-Id: Ifd96e409fb3049210fe9626b97997b08cea6f94c
JIRA: BGSOFUIPIRIN-3785