[FIX] sap.ui.layout.Grid: Wrapping of Grid children in IE is fixed
Problem: Grid children wrap too early in IE browser because their width
was not calculated with the correct calc function, taking in
condiseration the available space, number of items and their left and
right margins. The issue is introduced with an old change, which main
purpose was to floor the percentage width value of Grid children for IE
browser but has skipped the call to calc function, which leads to the
above mentioned problem.

Solution: Using the calc function now to calculate the width of Grid
children, keeping the old logic for flooring the percentage value as
this is related to a previous change, which has fixed a problem with
not rounded percentage values in IE.

Change-Id: I767ae1b998aca6db1875150e7fb4efe7702d835c
BCP: 1970156801