[INTERNAL][FIX] Publishing with rta works in case of url based app variants
If an app variant by a url is given, there is no appdescriptor variant
file existing which can be found. With this change in such a case, the
file is not requested.
The request of the file results in a broken promise chain which stops
the publishing of adaptations (shipment / key user adaptations in the
run time).

Change-Id: Idf0eb3fb078803ac850e8a7c09cb6b5c000ad917
BCP: 1980019097
CR-Id: 002075125800001050582019