[FIX] sap.uxap.ObjectPageLayout: Attach resize handler after rendering
Problem: When using standart Snippix template with synchronous rendering
of views, without awaiting for the Core to get initialized, the resize
handler for screen size update of ObjectPageLayout is never registered
and never called.

Solution: Attach the screen size update handler in onAfterRendering
method so that even in Snippix samples ObjectPageLayout will work as
expected.

Change-Id: Ide980d554ff67663cb65158aef353f23c1a9c1fe
BCP: 1980069920