[INTERNAL] sap.ui.table: avoid unintended use of sap.m helper in tests
The new module CreationRow depends on sap.m controls. Tests that require
that module, indirectly execute sap/m/library and thereby install the
sap.m implementation of the sap.ui.table.TableHelper. This deactivates
the dummy table helper of TableQUnitUtils and might lead to sync
requests when sap.m is not preloaded.

Change-Id: I4ef9beeaa2ef811bf2a7dff5684fec45cf3b5433