[FIX] sap.uxap.AnchorBar: Updated resize behavior
Problem description:
AnchorBar would sometimes appear incorrect when going to a Phone
breakpoint.

Root cause:
A function called in the AnchorBar "_adjustSize" function, would return
incorrect value. The function name is "_getCurrentMediaContainerRange"
and it would return incorrect contextual settings width.

Solution:
Instead of depending on contextual settings width, we are now measuring
the offsetWidth of the dom reference of the AnchorBar itself.

Change-Id: Ibacbc046c0d54d70612ae98d61df3f667fb73f07
BCP: 1980015066