[FIX] Core: Ensure firing the ThemeChanged event when loading libs
By passing "true" to the ThemeCheck it might happen that there is no
ThemeChanged event fired in case the check succeeds in the first run.

This flag should only be passed from Core#init where the ThemeCheck is
initialized and initially called as there are anyway no listeners for
the event, yet.

Change-Id: Ib330531130c762f43e3f043a60a262fac8a1e0cd
BCP: 0020751294 0000113165 2019