[FIX] sap.m.MultiInput: Use list in delete mode for N-more popover
The single source of truth for the content of the MultiInput are
the contained tokens. The list items in the N-more popover should
represent their actual state - meaning that only existing tokens
should be represented with a list item in the N-more popover.

In the previos implementation, a multiselect list was used. The issue
came from the fact that when a list item is deselected from the popover,
a new state was introduced which does not exist in the context of the token.
In that state, the list was represeting a token which is already removed
from the MultiInput.

On the other hand, once the user tried to add a token by selecting a previously
deselected list item from the N-more popover, the control notified the app
for the change. At this point the app could change the MultiInput content
(by altering the tokens aggregation, updating binding, etc.), which might lead
to synchronization issues between the tokens and the items, represented in the
N-more list. To prevent this, we are now changing the interaction of the list
items so that it corresponds to the interaction options with the actual tokens.
Now, tokens can only be deleted from the N-more popover and not added.

Change-Id: I53c66b1d392ca96e7df30fee12103f4cc4a74a53
BCP: 1970156490