[INTERNAL] m.SplitButton: Hidden tooltip label removed
The tooltip was previously set as a hidden label, which was
then used as a reference in aria-labelledby. This was done
for IE, in order for the tooltip to be read out. Since we
are switching to Chrome as a platform for testing and such
hack isn't needed there, we can safely remove it. Otherwise
the screen reader will just announce the tooltip twice.

Change-Id: I0a99eea507d3195ef94554e2b77959cad6799f6f
JIRA: BGSOFUIBALKAN-4693