[INTERNAL] m.ObjectHeader: Focus target changed
Until now when OH had active icon, the focus fell on the "-titleIcon"
div, which wrapped the actual icon. This was ok for IE, since the
icon's tooltip was read out, but for Chrome - it led to duplicate
readings. Besides that, half of the ACC attributes were placed on the
wrapper and the other half - on the actual icon. This is why now
the focusable element will be the actual icon and all attributes will
be placed on it, rather than the wrapper.

In addition to this, the element with overflow: hidden is now the "-title"
wrapper, rather than the "-titlediv". This way the active icon no longer
needs to have it's outline offset by -1px.

Change-Id: I3f1dbe1c5e8258d59dcb01932a546d4fcdeaf8f9
JIRA: BGSOFUIBALKAN-4693