[INTERNAL] sap.uxap.ObjectPageLayout: Title is measured without cloning
Problem: A clone of the snapped/expanded title DOM node was made each
time the header gets snapped/expanded and rendered on a hidden place in
order to be measured. However, this solution led to some performance
issues of requesting the image in the title on each snap/expand.

Solution: Now, instead of cloning the DOM node, a CSS stickied class is
only toggled and the title gets measured in both states. This way the
image is not requested because there is no cloning.

Change-Id: I10b1f1f467fa2bfdabebf729fd624a1ab250bbc2
JIRA: BGSOFUIPIRIN-3689