[FIX] sap.m.Select: Demo Kit samples binding fix
The problem:
In Demo Kit samples of sap.m.Select, if you open the select list by:
* Clicking on the arrow or icon
or
* Pressing Space and keyboard arrow down key
to change the selected item, the actual selected item changes and
then the control flashes and the previously selected item is again
selected preventing the user from changing the selected item.

The root cause:
In this sample, the control is bound to a list with keys and the
"selectedKey" is also bound to the initial selected item from that same
list. Upon navigation to the next item with the arrow key, we update the
"selectedKey" value as it is initiated by interaction.
The user interaction is bound to the second item in the list, changing
its "key" in the process as in this sample we are in TwoWay binding mode.
The final result is a list with two identical keys and our contract for
"selectedKey" is that we always select the first "item" in the list with
matching "key". Navigating up from the selected item is working as
expected, so the next selected item in this direction always precedes the
initially selected item in the list.
This is a wrong way of binding a control and the control behaves as
expected and in this case the samples need to be updated.

The solution:
Introduced a separate entry in the sample models which is then used to
bind the selectedKey property.

Note:
In the "Select" sample, all three controls are bound to a single property
which now updates all controls on the page upon changing the selection
of one.

Change-Id: I78d9df8b0e070cc716c5609fcbb2e5fe3ebcd210
BCP: 1980103137