[FIX] sap.m.GenericTag: Fixed onkeydown with SPACE key
Problem:
When the GenericTag is in a scrollable page and Space key
is pressed while the control is focused, the page was
scrolled which is incorrect behavior.

Fix:
Now, if the onkeydown event is triggered with SPACE key, GenericTag
goes only in active state and the event is prevented.

Change-Id: I3ffede6fbe08552481f5ec357dad4f6c4e8b361c
BCP: 1980029923