[INTERNAL] sap.m.InputBase: Layout adjustments and min width
JIRA: BGSOFUIRILA-2310
BCP: 002075129500001574612019

This change removes all styling and coding related to flex, which
was added in the first part of the input controls alignment. The
reason for this change are several issues regaring IE and CSS flex.

Now, all calculations are based on the min width principle.
Input controls should have a min width equal to the sum of
the min width of the inner input, the icon(s) min width,
borders and min width of a tokenizer with tokens (if present).

Example:
An editable MultiInput control with tokens in Cozy mode
has the following min width:
3rem(inner input) + 2.375rem(icon) + 3.25rem(tokenizer) + borders
Change-Id: Ib8e01430cb2a512f398fa600386554d5003ff0a7
CR-Id: 002075125900001499562019