[FIX] sap.uxap.ObjectPageSection: Showing empty title in RTA is fixed
Problem: When ObjectPageSection has only one SubSection, the title of
the SubSection is shown in the AnchorBar. However, if the only one
SubSection does not have title, the title of the Section is shown. If in
RTA scenario you delete the whole section and its SubSection does
not have a title, when you try to reveal back the Section, the
suggestion Dialog does not show any title.

Solution: Now upon revealing, the title of the only one SubSection is
returned only if it is not empty. If it is empty, the title of the
Section is returned and shown in the suggestion Dialog.

Change-Id: Id4f3cc7c114a4ddee8f966399ceaf1e368e7babf
BCP: 1970093334