[FIX] sap.m.SplitContainer: Hide master page in "HideMode" on mobile
Problem: When the SplitContainer is used in "HideMode" on mobile
the master page was shown initially.

Solution: The initial page of oMasterNav is set to the first detail
page or if initial detail page is set.

BCP: 002028376500005408012018
CR-Id: 002075125800001603322019
(cherry picked from commit 41bb019994a316da4fcb8cf4cfda2b0662497a5b)

Change-Id: I5d6fc1f238d111dea126f13e75039a0aade996b1