[INTERNAL] All libraries: fix lists of controls, elements and types
In some libs, the library metadata (maintained as parameter to the
initLibrary call) was not consistent with the public documentation or
the inheritance hierarchy of the listed entities.

This change fixes the following mismatches:

- removes entities from the lists that neither inherit from control
  nor element
- add types, elements or controls to the corresponding list that had
  been missing but which are documented as public

Entities that are private or that are not a module of their own are kept
in the lists for now. If such entities have been missing in the lists,
they have NOT been added. Whether they should/must/must not be listed
still has to be decided.

The cleanup is a preparation before enabling the generic tests for a
broader range of libs.

Change-Id: I89dfdcc4a56227df21a36a8ac92fd19a66c35ff3