[FIX] sap.m.Toolbar: Fiori 3 background for design Auto is fixed
Problem: When the design property of Toolbar and OverflowToolbar is set
to Auto, their background should be the same as the background of their
parent control. However, in Fiori 3 theme currently it is always Solid
(white color).

Solution: The background color of the Toolbar and OverflowToolbar is now
transparent, when their design property is set to Auto.

Change-Id: I3d08488ce0e347c8ad1ee06cf15803f0a4843c40
BCP: 1980134339