[INTERNAL][FIX] TestStarter: suite() should propagate errors
When an error occurred while creating the suite of test pages, the
generic test starter rendered the error on the screen but did not
propagate the error. This was well-suited for a human user, but did not
make the testrunner aware of the error.

With this change, such an error is now rendered AND propagated to the
caller. It is still up to the caller to react accordingly.

Change-Id: I0cf7e947442d8d81b30ac06aaec43b16023eebf4