[INTERNAL][FIX] sap.m.ComparisonPattern: All properties are now displayed
Problem: On large screens, the number of properties of the compared 
items are not enough to make the title of the DynamicPage to go into
snapped state.

Solution: All properties are displayed now, so the snapping position can
be reached even on large screens. This way the functionality of
snapping/expanding of the title of the DynamicPage on scroll is also
demonstrated in the Sample.

Change-Id: Ia67a1975ab919c4862199e7561f4b1fc91b0e9a4
JIRA: BGSOFUIPIRIN-3337