[INTERNAL] ManagedObjectModel: Paging in JSON model stops a size limit
Note:
If a table or a list base is bound to an aggregation of a managed object
that is not growing enabled (like pages/assetPages) then the paging of
that table though stops at the size limit of the inner client model.

Note 2:
We adapted a change I6d38b14ec3b17affc13dbd320e57308f6ff67afa that
already addressed paging and which was pretty good working for OData
model but not for client model. We missed that gap as our unit tests
never reached the size limit of the model.
When adapting this we realized that we where to force a paging/update
in the inner aggregation.
Now to be on a save side for controls that have their own update method
we made the aggregation update accessor available.
The complete change is surrounded with 4 unit tests.
See for snippix in the internal message.

Change-Id: I3e44926e92629260b86d49c0152bee1a8c550a5a
BCP: 1980140825