[INTERNAL][FIX] sap.m.ProgressIndicator: Text margins/paddings aligned
BCP: 1980133537

In the minimum size of the ProgressIndicator, the text was too far in
the rounded corner of the bar.

We synced with VD and decided to apply margins on both sides of the text
in Fiori 3 (spec is updated).

Since the problem is present in Belize also, the left padding of the
problematic case is also set to be equal to the right one.

High contrast themes do not require adjustments, since there is an icon
before the text with a margin already applied.

Change-Id: I880d3ec5dde3432d0422e3f09e09f5164d913940