[INTERNAL] RM: style and class writing adaptations
 - Aligns style writing with the CSSStyleDeclaration spec. e.g.
   style="color:black" vs style="color: black;"
 - Additional duplicate class name check is removed. Having duplicate
class names in one HTMLElement is not necessary but this should be
ensured by the consumer. Since jQuery#removeClass, jQuery#addClass and
DOMTokenList take care of them correctly this additional effort is
removed.
 - Adding custom style classes on before rendering hook must be taken
into account by the RenderManager.
 - Empty style value should be ignored by the addStyle otherwise
consumer must check with "if" to ignore empty control properties.
 - Instead of pushing many sub strings to the RM buffer, we are using
string concatenation.
 - voidEnd bug is fixed.
 - Unit tests are adapted.

Change-Id: If290e6bb5e25ba1076f30535e300c612e9513765