[INTERNAL] sap.ui.integration.host.HostConfiguration: AnalyticalContent is no longer a dependency
Issue:
When the local openui5 tests are run the sap.ui.core testsuite.generic.qunit.html
was failing because sap.viz library is not present in openui5. (HostConfiguration control
is loaded but it has a hard-coded dependency to AnalyticalContent which has viz dependencies)

Solution:
AnalyticalContent dependency is removed from the HostConfiguration.

Change-Id: Iafc07e6ee472c2afcc5ae01938ff7b6c67c51fea