[INTERNAL] sap.uxap.ObjectPageLayout: Background Image mode is used
Using the Backround mode of the image in order to render span with
background-image, instead of img tag with src. This way, the image is
not getting requested on each snap/expand of the Header or on each
scroll, when we clone the Header in order to measure it.

Change-Id: I449fcc28c53a70f59efe103abbed7adf7f7cd0a9
JIRA: BGSOFUIPIRIN-3831