[FIX] P13nDialog: navigation displaying improvements
When there are multiple panels in the P13nDialog, but only of the same
type, the P13nDialog did not handle the navigation displaying of the
panels correctly.

Now the P13nDialog is displaying several panels of the same type as
different panels, even if they are all of the same type.

BCP: 1980147525
Change-Id: I32c3c8fbabb5e75bd780d6673266dfae406ecd39