[INTERNAL] trace/FESR: Send Passport action via client ID field
To allow easier correlation between Passports and FESR a key needs
to be sent with the FESR to identify the according Passports. Therefore
the unused client ID field is taken to provide the action sent with
the Passport. Together with the root context id this provides a way
to correlate Passports and FESR.

Change-Id: I937af472d433d8ebbb5692e5c23b3ee810b5c8d5
JIRA: CPOUIFPHOENIXCORE-2778