[FEATURE] Send FESR via Beacon API
As the piggyback procedure of transporting FESR header has some reliability
issues, a new procedure is established by leveraging the Beacon API of the
browser. With this change it is possible to pass an URL to the FESR module
to which the FESR will be sent instead.
This is done by buffering them to a certain threshold (default=10 entries)
and sending them via navigator.sendBeacon. Furthermore the close event of
the window will cause the buffer to send all appended entries so far, in
order to not lose any.

JIRA: CPOUIFPHOENIXCORE-2772
Change-Id: Ica5c094a5dd39422606fcdd40b27716cb9a80330