[FEATURE] sap.m.Select: Control can now be both disabled and focusable
The new editable property is of type boolean. Its default value is true.

If editable is set to false, the interaction with the control is
disabled.
The control remains focusable by design. Its background and
border are changed to indicate that it's read-only.

Additionally the "Select" sample in the "Demo Kit" is updated.
Two buttons ("Toggle Enabled", "Toggle Editable") are added, so it would
be easier to distinguish the difference of "Enabled" and "Editable"
properties.

Change-Id: I76c8c45ff9a12777d95033543e14efbda42b2776
JIRA: BGSOFUIPIRIN-1413