[INTERNAL] performance/XHRInterceptor
A problem occured with registration of callbacks, as a previously
registered callback may not be overwritten properly. With this change
the structure of the internal registry is changed so this cannot
happen any longer.
Furthermore the return value of the delete method has been slightly
changed to comply to the behaviour of the delete operator.

Change-Id: I46b8d3ec170c0cb805e58f5c8533e11f3abf6452