[FEATURE] sap.f.ShellBar: Two main types of behavior now supported
With “showMenuButton” property to “true”, we have a hamburger menu
button and plain title, which is hidden on mobile view.

With “showMenuButton” set to “false”, the control shows the title as
part of a mega menu, which is configurable by the app developer. It
also is hidden on mobile.

Change-Id: I9ceb2da26923f897b32b496f06282e3244770a38
JIRA: BGSOFUIPIRIN-3832