[FIX] sap.m.DynamicPageTitle: Fixed incorrect display of aggregation
Problem description:
SnappedTitleOnMobile aggregation of the sap.m.DynamicPageTitle is not
displaying in the correct position coordinates on some phone devices.

Root cause:
One of the DOM siblings of the SnappedTitleOnMobile wrapper is
incorrectly positioned relative, instead of absolute, when the
sapUiHidden class is applied. This would make the problematic sibling
to take space when it isn't supposed to, resulting in pushing down
the SnappedTitleOnMobile wrapper.

Solution:
We force the problematic sibling to be absolutely positioned, when the
sapUiHidden class is applied.

Change-Id: Id38c1bee6d05b2368b3aa9956412f9e983fcaa84
BCP: 1980146670