[FIX] sap.m.Select: Fixed picker height in IE11
Problem:
The sap.m.SelectList of sap.m.Select is not displayed in IE11.

Root cause:
The flex content of sap.ui.layout.SimpleFixFlex has a CSS
style "flex:1". However IE 11 requires a unit to be added
to the third argument, the flex-basis property

Solution:
Change the flex style of the flex content of
sap.ui.layout.SimpleFixFlex

CR-Id: 002075125800001876702019
Change-Id: I0a0376757e53dded444961f58b46c48121eb75a6
BCP: 1980165515