[FIX] sap.m.TextArea: Growing property is now corrected
Previously, when a growing TextArea was inside a container
with scroll, typing in the TextArea would cause jumping of
the parent container with scroll. The issue is present in IE
and partially in Chrome browsers and was caused by the
combination of auto height and resetting of the height of the
textarea HTML element.
Now, a mirror div is created when the growing property is true.
It stretches the wrapper container when needed and without
causing 'jumps'.

BCP: 1940164068

Change-Id: Ia793e108d0d8a16436f74fe7ab657c68059c4fa8
CR-Id: 002075125900001523312019