[FIX] ManagedObjectModel: Inform when parts of a property change
Note:
Within the managed object model it is possible to completely bind
against the API of a managed object. If these properties are array or
object like binding paths tend to look like /objectValue/value or
/objectArray/0/values. In any of these cases just setting the value
inside the part of that property will not be detecting any change as the
reference is still the same.
In order to fix this the managed object model fetches the complete
object/array that corresponds to the property and clones it.
The clone is then changed accordingly and the complete new value is fed
into the API of the managed object which then changes the complete
property, informs the binding and triggers the remaining processing.

Change-Id: I216da74f14fb5b03817911e8b23c0e05589a3f52
BCP: 1980169334