[INTERNAL][FIX] sap.m.Dialog: Content width in Mozilla/IE/Edge fixed
Issue:
When there is vertical scrolling in the dialog in IE/Mozilla/Edge
and the content is smaller than the min-width of the dialog
there is a gap on the right side.

Solution:
Only set width auto when the content is bigger than the min-width
of the dialog in order to make it grow.

BCP: 1980156441
Change-Id: I3729b71b994c93277e5cee64eca2eacc0acdacb9
(cherry picked from commit 5b057f9f01d5fc85f45d4819585b93421bd87efa)
CR-Id: 002075125900005888192019

BCP: 1980420226